Revert "Clear delivery bucket's module-boundary violations, graduate it to enforced"
This reverts commit cb91f5887d43e1dccf3ac9053fa516a6dc19229f.
