+++ apps/api/src/agent-surface/mcp-source-stage-tools.ts
@@ -3,7 +3,7 @@ import { AGENT_CHANNEL_ROUTES } from '@gamedevpl/contract';
 import { canonicalAppBaseUrl } from '../platform/canonical-app-url.js';
 import { DEFAULT_UPLOAD_URL_TTL_SECONDS, mintUploadToken, uploadCurlCommand } from './agent-upload-token.js';
 import { assertDeliverableSourcePath, InvalidUploadError } from '../delivery/games-store.js';
-import { decodeRasterSourceContent, encodeRasterSourceContent, isRasterSourcePath } from '../platform/raster-source.js';
+import { decodeRasterSourceContent, encodeRasterSourceContent, isRasterSourcePath } from '../catalog/raster-assets.js';
 import { decodeCanonicalBase64Utf8, InvalidBase64Error } from '../platform/canonical-base64.js';
 import { largeSourceFileHint, moduleSizeWarnings } from '../creation/module-size.js';
 import { gameManifestHint } from './game-manifest-hint.js';
