+++ apps/api/src/catalog/index-html-generator.ts
@@ -3,10 +3,20 @@
 // Contract and goldens: docs/how-to-play-plan.md
 
 import type { Locale } from '@gamedevpl/contract';
-import { hasPlayableHowToPlay, type HowToPlay } from '../platform/how-to-play.js';
 
-export type { HowToPlay };
-export { hasPlayableHowToPlay };
+export interface HowToPlay {
+  controls?: Array<{
+    keys: string | { en: string; pl: string };
+    action: { en: string; pl: string };
+  }>;
+  goal: { en: string; pl: string };
+  scoring?: { en: string; pl: string };
+  mode?: { en: string; pl: string };
+  hint: { en: string; pl: string };
+  sound?: { en: string; pl: string } | false;
+  playAgain?: { en: string; pl: string } | false;
+  touch?: { en: string; pl: string } | false;
+}
 
 export interface Canvas {
   width?: number;
@@ -26,6 +36,24 @@ export interface GameManifest {
 export interface GameSpec {
   title: string;
   [key: string]: unknown;
+}
+
+function isBilingualString(value: unknown): value is { en: string; pl: string } {
+  return (
+    typeof value === 'object' &&
+    value !== null &&
+    typeof (value as { en?: unknown }).en === 'string' &&
+    (value as { en: string }).en.length > 0 &&
+    typeof (value as { pl?: unknown }).pl === 'string' &&
+    (value as { pl: string }).pl.length > 0
+  );
+}
+
+// Shared by upload validation, generation, and staged-preview readiness — see docs.
+export function hasPlayableHowToPlay(howToPlay: unknown): howToPlay is HowToPlay {
+  if (typeof howToPlay !== 'object' || howToPlay === null) return false;
+  const candidate = howToPlay as { goal?: unknown; hint?: unknown };
+  return isBilingualString(candidate.goal) && isBilingualString(candidate.hint);
 }
 
 interface LegendRow {
