+++ apps/api/src/delivery/build-changelog.test.ts
@@ -9,7 +9,6 @@ import {
   resolveChangelogText,
   type ChangelogEvent,
 } from './build-changelog.js';
-import { isMcpPresenceEventText } from '../agent-surface/mcp-presence.js';
 import type { RecentBuild } from '@gamedevpl/contract';
 
 function event(partial: Partial<ChangelogEvent> & Pick<ChangelogEvent, 'kind' | 'text' | 'createdAt'>): ChangelogEvent {
@@ -28,36 +27,23 @@ function build(partial: Partial<RecentBuild> & Pick<RecentBuild, 'version' | 'cr
 describe('isChangelogWorthy', () => {
   it('keeps done and step sentences', () => {
     expect(
-      isChangelogWorthy(
-        event({ kind: 'done', text: 'Jump feels tighter.', createdAt: '2026-08-23T01:00:00.000Z' }),
-        isMcpPresenceEventText,
-      ),
+      isChangelogWorthy(event({ kind: 'done', text: 'Jump feels tighter.', createdAt: '2026-08-23T01:00:00.000Z' })),
     ).toBe(true);
     expect(
       isChangelogWorthy(
         event({ kind: 'step', text: 'Wiring the landing HUD.', createdAt: '2026-08-23T01:00:00.000Z' }),
-        isMcpPresenceEventText,
       ),
     ).toBe(true);
   });
 
   it('drops blocked, asking, empty, and leftover presence rows', () => {
     expect(
-      isChangelogWorthy(
-        event({ kind: 'blocked', text: 'Need a typecheck.', createdAt: '2026-08-23T01:00:00.000Z' }),
-        isMcpPresenceEventText,
-      ),
+      isChangelogWorthy(event({ kind: 'blocked', text: 'Need a typecheck.', createdAt: '2026-08-23T01:00:00.000Z' })),
     ).toBe(false);
-    expect(
-      isChangelogWorthy(
-        event({ kind: 'step', text: '   ', createdAt: '2026-08-23T01:00:00.000Z' }),
-        isMcpPresenceEventText,
-      ),
-    ).toBe(false);
+    expect(isChangelogWorthy(event({ kind: 'step', text: '   ', createdAt: '2026-08-23T01:00:00.000Z' }))).toBe(false);
     expect(
       isChangelogWorthy(
         event({ kind: 'step', text: 'Browsing the Creator Kit…', createdAt: '2026-08-01T01:00:00.000Z' }),
-        isMcpPresenceEventText,
       ),
     ).toBe(false);
   });
@@ -125,27 +111,17 @@ describe('pickChangelogEvent', () => {
   ];
 
   it('takes the done event in the version window', () => {
-    const picked = pickChangelogEvent('2026-08-23T02:00:00.000Z', undefined, events, isMcpPresenceEventText);
+    const picked = pickChangelogEvent('2026-08-23T02:00:00.000Z', undefined, events);
     expect(picked?.text).toBe('Sealed the publish.');
   });
 
   it('takes the newest step when the next version cuts off done', () => {
-    const picked = pickChangelogEvent(
-      '2026-08-23T02:00:00.000Z',
-      '2026-08-23T02:30:00.000Z',
-      events,
-      isMcpPresenceEventText,
-    );
+    const picked = pickChangelogEvent('2026-08-23T02:00:00.000Z', '2026-08-23T02:30:00.000Z', events);
     expect(picked?.text).toBe('Fixing collisions.');
   });
 
   it('falls back to the last step before the version when the window is empty', () => {
-    const picked = pickChangelogEvent(
-      '2026-08-23T01:40:00.000Z',
-      '2026-08-23T02:00:00.000Z',
-      events,
-      isMcpPresenceEventText,
-    );
+    const picked = pickChangelogEvent('2026-08-23T01:40:00.000Z', '2026-08-23T02:00:00.000Z', events);
     expect(picked?.text).toBe('Drafting the first playable.');
   });
 });
@@ -153,13 +129,10 @@ describe('pickChangelogEvent', () => {
 describe('pickLatestChangelogText', () => {
   it('returns the newest worthy sentence', () => {
     expect(
-      pickLatestChangelogText(
-        [
-          event({ kind: 'blocked', text: 'Need TRACE.', createdAt: '2026-08-23T02:00:00.000Z' }),
-          event({ kind: 'step', text: 'Tuning the jump.', createdAt: '2026-08-23T01:00:00.000Z' }),
-        ],
-        isMcpPresenceEventText,
-      ),
+      pickLatestChangelogText([
+        event({ kind: 'blocked', text: 'Need TRACE.', createdAt: '2026-08-23T02:00:00.000Z' }),
+        event({ kind: 'step', text: 'Tuning the jump.', createdAt: '2026-08-23T01:00:00.000Z' }),
+      ]),
     ).toBe('Tuning the jump.');
   });
 });
@@ -180,7 +153,7 @@ describe('applyChangelogSummaries', () => {
       ],
     ]);
 
-    expect(applyChangelogSummaries(builds, events, isMcpPresenceEventText).map((item) => item.summary)).toEqual([
+    expect(applyChangelogSummaries(builds, events).map((item) => item.summary)).toEqual([
       'Sealed the publish.',
       'Drafting the first playable.',
     ]);
@@ -192,7 +165,6 @@ describe('hydrateRecentBuildSummaries', () => {
     const seen: number[] = [];
     const hydrated = await hydrateRecentBuildSummaries({
       locale: 'en',
-      isPresenceEventText: isMcpPresenceEventText,
       builds: [
         build({ version: 'v2', createdAt: '2026-08-23T02:00:00.000Z', jobId: 10, summary: 'Kept.' }),
         build({ version: 'v1', createdAt: '2026-08-23T01:00:00.000Z', jobId: 10 }),
