+++ apps/api/src/delivery/build-changelog.ts
@@ -1,6 +1,7 @@
 // Sentence changelog for a version, from the round's own events.
 
 import type { RecentBuild } from '@gamedevpl/contract';
+import { isMcpPresenceEventText } from '../agent-surface/mcp-presence.js';
 
 const CHANGELOG_KINDS = new Set(['done', 'step']);
 
@@ -12,14 +13,10 @@ export type ChangelogEvent = {
   createdAt: string;
 };
 
-// N1: injected so this module has no value-level agent-surface import.
-export function isChangelogWorthy(
-  event: ChangelogEvent,
-  isPresenceEventText: (text: string, createdAt?: string) => boolean,
-): boolean {
+export function isChangelogWorthy(event: ChangelogEvent): boolean {
   if (!CHANGELOG_KINDS.has(event.kind)) return false;
   if (!event.text.trim()) return false;
-  return !isPresenceEventText(event.text, event.createdAt);
+  return !isMcpPresenceEventText(event.text, event.createdAt);
 }
 
 export function resolveChangelogText(event: ChangelogEvent, locale?: string): string {
@@ -46,7 +43,6 @@ export function pickChangelogEvent(
   createdAt: string,
   nextCreatedAt: string | undefined,
   events: readonly ChangelogEvent[],
-  isPresenceEventText: (text: string, createdAt?: string) => boolean,
 ): ChangelogEvent | undefined {
   const start = Date.parse(createdAt);
   if (!Number.isFinite(start)) return undefined;
@@ -56,7 +52,7 @@ export function pickChangelogEvent(
   const inWindow: ChangelogEvent[] = [];
   const before: ChangelogEvent[] = [];
   for (const event of events) {
-    if (!isChangelogWorthy(event, isPresenceEventText)) continue;
+    if (!isChangelogWorthy(event)) continue;
     const at = Date.parse(event.createdAt);
     if (!Number.isFinite(at)) continue;
     if (at >= start && at < end) inWindow.push(event);
@@ -68,19 +64,14 @@ export function pickChangelogEvent(
   return preferDone(inWindow) ?? preferDone(before);
 }
 
-export function pickLatestChangelogText(
-  events: readonly ChangelogEvent[],
-  isPresenceEventText: (text: string, createdAt?: string) => boolean,
-  locale?: string,
-): string | undefined {
-  const event = events.find((item) => isChangelogWorthy(item, isPresenceEventText));
+export function pickLatestChangelogText(events: readonly ChangelogEvent[], locale?: string): string | undefined {
+  const event = events.find((item) => isChangelogWorthy(item));
   return event ? resolveChangelogText(event, locale) : undefined;
 }
 
 export function applyChangelogSummaries(
   builds: RecentBuild[],
   eventsByIssue: Map<number, readonly ChangelogEvent[]>,
-  isPresenceEventText: (text: string, createdAt?: string) => boolean,
   locale?: string,
 ): RecentBuild[] {
   const byIssue = new Map<number, RecentBuild[]>();
@@ -97,7 +88,7 @@ export function applyChangelogSummaries(
     const oldestFirst = [...group].sort((a, b) => a.createdAt.localeCompare(b.createdAt));
     for (let i = 0; i < oldestFirst.length; i++) {
       const build = oldestFirst[i]!;
-      const event = pickChangelogEvent(build.createdAt, oldestFirst[i + 1]?.createdAt, events, isPresenceEventText);
+      const event = pickChangelogEvent(build.createdAt, oldestFirst[i + 1]?.createdAt, events);
       const text = resolveBuildSummary(build.summary, event, locale);
       if (text) resolved.set(build.version, text);
     }
@@ -114,12 +105,11 @@ export async function hydrateRecentBuildSummaries(input: {
   builds: RecentBuild[];
   locale?: string;
   loadEvents: (jobId: number) => Promise<readonly ChangelogEvent[]>;
-  isPresenceEventText: (text: string, createdAt?: string) => boolean;
 }): Promise<RecentBuild[]> {
   const ids = [
     ...new Set(input.builds.map((build) => build.jobId).filter((id): id is number => typeof id === 'number')),
   ];
   if (ids.length === 0) return input.builds;
   const entries = await Promise.all(ids.map(async (id) => [id, await input.loadEvents(id)] as const));
-  return applyChangelogSummaries(input.builds, new Map(entries), input.isPresenceEventText, input.locale);
+  return applyChangelogSummaries(input.builds, new Map(entries), input.locale);
 }
