+++ eslint-rules/comment-prose-baseline.json
@@ -12,7 +12,7 @@
     "apps/api/scripts/remix-lane-judge.ts": 394,
     "apps/api/scripts/remix-lane-probe.ts": 221,
     "apps/api/scripts/remix-lane-sweep.ts": 313,
-    "apps/api/scripts/run-gate.ts": 890,
+    "apps/api/scripts/run-gate.ts": 692,
     "apps/api/scripts/slug-backfill.ts": 296,
     "apps/api/scripts/verify-erase-signals.ts": 217,
     "apps/api/src/agent-build-media.test.ts": 667,
@@ -74,6 +74,7 @@
     "apps/api/src/catalog/github-client.ts": 4294,
     "apps/api/src/catalog/local-games-repo.test.ts": 75,
     "apps/api/src/catalog/local-games-repo.ts": 373,
+    "apps/api/src/catalog/music-tracks.ts": 154,
     "apps/api/src/catalog/published-slugs.test.ts": 31,
     "apps/api/src/catalog/published-slugs.ts": 558,
     "apps/api/src/catalog/recommend.test.ts": 14,
@@ -261,7 +262,6 @@
     "apps/api/src/platform/moderation-terms.ts": 236,
     "apps/api/src/platform/moderation.test.ts": 99,
     "apps/api/src/platform/moderation.ts": 336,
-    "apps/api/src/platform/music-tracks.ts": 154,
     "apps/api/src/platform/oauth-as.test.ts": 340,
     "apps/api/src/platform/oauth-as.ts": 471,
     "apps/api/src/platform/oauth-page-chrome.ts": 255,
