+++ eslint-rules/module-boundary-map.mjs
@@ -31,7 +31,7 @@ export const MODULE_BUCKETS = [
  * `eslint .` pass; every other bucket stays warn-only via `npm run module-boundary` until
  * its own turn. Append to this list, never edit its enforcement elsewhere.
  */
-export const ENFORCED_BUCKETS = ['telemetry', 'catalog', 'realtime', 'notifications', 'community', 'delivery'];
+export const ENFORCED_BUCKETS = ['telemetry', 'catalog', 'realtime', 'notifications', 'community'];
 
 const DEFAULT_BUCKET = 'platform';
 
@@ -113,9 +113,6 @@ const FILE_BUCKET = {
   'games-repo-contract': 'platform',
   'games-repo-bundle-bytes': 'platform',
   'raster-contract': 'platform',
-  'raster-source': 'platform',
-  'music-tracks': 'platform',
-  'how-to-play': 'platform',
   'kit-registry': 'platform',
   'kit-window': 'platform',
   'round-base-version': 'platform',
@@ -294,6 +291,7 @@ const FILE_BUCKET = {
   'admin-game-routes': 'catalog',
   'game-play-route': 'catalog',
   'slug-resolver': 'catalog',
+  'music-tracks': 'catalog',
   'catalog-enricher': 'catalog',
   'catalog-indexer': 'catalog',
   'catalog-vector-index': 'catalog',
