+++ eslint-rules/module-size-baseline.json
@@ -17,7 +17,7 @@
     "apps/api/scripts/remix-lane-judge.ts": 216,
     "apps/api/scripts/remix-lane-probe.ts": 126,
     "apps/api/scripts/remix-lane-sweep.ts": 289,
-    "apps/api/scripts/run-gate.ts": 249,
+    "apps/api/scripts/run-gate.ts": 212,
     "apps/api/scripts/slug-backfill.ts": 95,
     "apps/api/scripts/verify-erase-signals.ts": 60,
     "apps/api/src/__tests__/index-html-contract.test.ts": 25,
@@ -33,7 +33,7 @@
     "apps/api/src/agent-surface/agent-build-brief.ts": 52,
     "apps/api/src/agent-surface/agent-build-examples.ts": 62,
     "apps/api/src/agent-surface/agent-channel.test.ts": 3408,
-    "apps/api/src/agent-surface/agent-channel.ts": 2292,
+    "apps/api/src/agent-surface/agent-channel.ts": 2284,
     "apps/api/src/agent-surface/agent-creator-key-resolve.test.ts": 114,
     "apps/api/src/agent-surface/agent-creator-key-resolve.ts": 170,
     "apps/api/src/agent-surface/agent-creator-key.test.ts": 133,
@@ -121,6 +121,8 @@
     "apps/api/src/catalog/index-html-generator.ts": 244,
     "apps/api/src/catalog/local-games-repo.test.ts": 158,
     "apps/api/src/catalog/local-games-repo.ts": 212,
+    "apps/api/src/catalog/music-tracks.test.ts": 74,
+    "apps/api/src/catalog/music-tracks.ts": 224,
     "apps/api/src/catalog/published-slugs.test.ts": 175,
     "apps/api/src/catalog/published-slugs.ts": 149,
     "apps/api/src/catalog/recommend.test.ts": 167,
@@ -186,11 +188,11 @@
     "apps/api/src/creation/creation-limits.test.ts": 364,
     "apps/api/src/creation/creation-limits.ts": 508,
     "apps/api/src/creation/creator-code.test.ts": 1398,
-    "apps/api/src/creation/creator-code.ts": 1336,
+    "apps/api/src/creation/creator-code.ts": 1330,
     "apps/api/src/creation/creator-profile-routes.test.ts": 279,
     "apps/api/src/creation/creator-profile-routes.ts": 317,
     "apps/api/src/creation/creator-studio.test.ts": 734,
-    "apps/api/src/creation/creator-studio.ts": 531,
+    "apps/api/src/creation/creator-studio.ts": 529,
     "apps/api/src/creation/dispatch-reaper.test.ts": 77,
     "apps/api/src/creation/dispatch-reaper.ts": 98,
     "apps/api/src/creation/editor-assist.test.ts": 122,
@@ -256,13 +258,13 @@
     "apps/api/src/creation/typecheck-preflight.test.ts": 199,
     "apps/api/src/creation/typecheck-preflight.ts": 246,
     "apps/api/src/creator-studio-workspace.test.ts": 321,
-    "apps/api/src/delivery/build-changelog.test.ts": 213,
-    "apps/api/src/delivery/build-changelog.ts": 125,
+    "apps/api/src/delivery/build-changelog.test.ts": 185,
+    "apps/api/src/delivery/build-changelog.ts": 115,
     "apps/api/src/delivery/build-preview-limits.ts": 2,
     "apps/api/src/delivery/build-prompt.test.ts": 171,
     "apps/api/src/delivery/build-prompt.ts": 186,
-    "apps/api/src/delivery/build-transcript.test.ts": 339,
-    "apps/api/src/delivery/build-transcript.ts": 162,
+    "apps/api/src/delivery/build-transcript.test.ts": 334,
+    "apps/api/src/delivery/build-transcript.ts": 160,
     "apps/api/src/delivery/games-store.test.ts": 1128,
     "apps/api/src/delivery/games-store.ts": 1398,
     "apps/api/src/delivery/gate-crash.test.ts": 180,
@@ -279,10 +281,10 @@
     "apps/api/src/delivery/gcs-sign.test.ts": 62,
     "apps/api/src/delivery/gcs-sign.ts": 193,
     "apps/api/src/delivery/recent-builds.ts": 50,
-    "apps/api/src/delivery/source-delivery.test.ts": 606,
-    "apps/api/src/delivery/source-delivery.ts": 540,
-    "apps/api/src/delivery/stage-hints.test.ts": 334,
-    "apps/api/src/delivery/stage-hints.ts": 186,
+    "apps/api/src/delivery/source-delivery.test.ts": 599,
+    "apps/api/src/delivery/source-delivery.ts": 536,
+    "apps/api/src/delivery/stage-hints.test.ts": 331,
+    "apps/api/src/delivery/stage-hints.ts": 179,
     "apps/api/src/delivery/staged-preview.test.ts": 692,
     "apps/api/src/delivery/staged-preview.ts": 614,
     "apps/api/src/delivery/workspace-archive.test.ts": 189,
@@ -377,8 +379,6 @@
     "apps/api/src/platform/moderation-terms.ts": 120,
     "apps/api/src/platform/moderation.test.ts": 228,
     "apps/api/src/platform/moderation.ts": 247,
-    "apps/api/src/platform/music-tracks.test.ts": 74,
-    "apps/api/src/platform/music-tracks.ts": 224,
     "apps/api/src/platform/oauth-as.test.ts": 953,
     "apps/api/src/platform/oauth-as.ts": 799,
     "apps/api/src/platform/oauth-page-chrome.ts": 82,
@@ -439,7 +439,7 @@
     "apps/api/src/store/firestore.ts": 1350,
     "apps/api/src/store/in-memory.ts": 1286,
     "apps/api/src/submissions.test.ts": 7368,
-    "apps/api/src/submissions.ts": 1700,
+    "apps/api/src/submissions.ts": 1693,
     "apps/api/src/tab-complete-deploy-flag.test.ts": 28,
     "apps/api/src/telemetry-trends.route.test.ts": 75,
     "apps/api/src/telemetry/chat-agent-metrics.test.ts": 53,
